move general_midi
