+++ src/core/general_midi.rs
@@ -1,0 +1,138 @@
+#[repr(u8)]
+#[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash)]
+pub enum GeneralMidi {
+    AcousticGrandPiano = 1,
+    BrightAcousticPiano = 2,
+    ElectricGrandPiano = 3,
+    HonkyTonkPiano = 4,
+    ElectricPiano1 = 5,
+    ElectricPiano2 = 6,
+    Harpsichord = 7,
+    Clavi = 8,
+    Celesta = 9,
+    Glockenspiel = 10,
+    MusicBox = 11,
+    Vibraphone = 12,
+    Marimba = 13,
+    Xylophone = 14,
+    TubularBells = 15,
+    Dulcimer = 16,
+    DrawbarOrgan = 17,
+    PercussiveOrgan = 18,
+    RockOrgan = 19,
+    ChurchOrgan = 20,
+    ReedOrgan = 21,
+    Accordion = 22,
+    Harmonica = 23,
+    TangoAccordion = 24,
+    AcousticGuitarNylon = 25,
+    AcousticGuitarSteel = 26,
+    ElectricGuitarJazz = 27,
+    ElectricGuitarClean = 28,
+    ElectricGuitarMuted = 29,
+    OverdrivenGuitar = 30,
+    DistortionGuitar = 31,
+    GuitarHarmonics = 32,
+    AcousticBass = 33,
+    ElectricBassFinger = 34,
+    ElectricBassPick = 35,
+    FretlessBass = 36,
+    SlapBass1 = 37,
+    SlapBass2 = 38,
+    SynthBass1 = 39,
+    SynthBass2 = 40,
+    Violin = 41,
+    Viola = 42,
+    Cello = 43,
+    Contrabass = 44,
+    TremoloStrings = 45,
+    PizzicatoStrings = 46,
+    OrchestralHarp = 47,
+    Timpani = 48,
+    StringEnsemble1 = 49,
+    StringEnsemble2 = 50,
+    SynthStrings1 = 51,
+    SynthStrings2 = 52,
+    ChoirAahs = 53,
+    VoiceOohs = 54,
+    SynthVoice = 55,
+    OrchestraHit = 56,
+    Trumpet = 57,
+    Trombone = 58,
+    Tuba = 59,
+    MutedTrumpet = 60,
+    FrenchHorn = 61,
+    BrassSection = 62,
+    SynthBrass1 = 63,
+    SynthBrass2 = 64,
+    SopranoSax = 65,
+    AltoSax = 66,
+    TenorSax = 67,
+    BaritoneSax = 68,
+    Oboe = 69,
+    EnglishHorn = 70,
+    Bassoon = 71,
+    Clarinet = 72,
+    Piccolo = 73,
+    Flute = 74,
+    Recorder = 75,
+    PanFlute = 76,
+    BlownBottle = 77,
+    Shakuhachi = 78,
+    Whistle = 79,
+    Ocarina = 80,
+    Lead1Square = 81,
+    Lead2Sawtooth = 82,
+    Lead3Calliope = 83,
+    Lead4Chiff = 84,
+    Lead5Charang = 85,
+    Lead6Voice = 86,
+    Lead7Fifths = 87,
+    Lead8BassPlusLead = 88,
+    Pad1Newage = 89,
+    Pad2Warm = 90,
+    Pad3Polysynth = 91,
+    Pad4Choir = 92,
+    Pad5Bowed = 93,
+    Pad6Metallic = 94,
+    Pad7Halo = 95,
+    Pad8Sweep = 96,
+    Fx1Rain = 97,
+    Fx2Soundtrack = 98,
+    Fx3Crystal = 99,
+    Fx4Atmosphere = 100,
+    Fx5Brightness = 101,
+    Fx6Goblins = 102,
+    Fx7Echoes = 103,
+    Fx8SciFi = 104,
+    Sitar = 105,
+    Banjo = 106,
+    Shamisen = 107,
+    Koto = 108,
+    Kalimba = 109,
+    Bagpipe = 110,
+    Fiddle = 111,
+    Shanai = 112,
+    TinkleBell = 113,
+    Agogo = 114,
+    SteelDrums = 115,
+    Woodblock = 116,
+    TaikoDrum = 117,
+    MelodicTom = 118,
+    SynthDrum = 119,
+    ReverseCymbal = 120,
+    GuitarFretNoise = 121,
+    BreathNoise = 122,
+    Seashore = 123,
+    BirdTweet = 124,
+    TelephoneRing = 125,
+    Helicopter = 126,
+    Applause = 127,
+    Gunshot = 128,
+}
+
+impl Into<u8> for GeneralMidi {
+    fn into(self) -> u8 {
+        self as u8
+    }
+}
